reeds should work with the new merge branch

+++ gromosXX/src/check/test_system/test_reeds/submit.sh
@@ -1,20 +1,23 @@
+#Load Gromos Settings
+. ../gromos_settings.sh
 
-. ../../gromos_settings.sh
 
-for x in $( ls ${GROMSRC}/testbuild* -d);
+for x in $( ls ${GROMDIR}/testbuild* -d);
 do 
-	echo ${x};
-	binDIR="${x}/bin"
-	dirP=$(basename $x)
-	dirP=${dirP/build_}
+	if [[ ${x} == *"testbuild"*"mpi"* ]];
+	then
+		echo ${x};
+		binDIR="${x}/bin"
+		dirP=$(basename $x)
+		dirP=${dirP/build}
 
-	echo ${dirP}
-	rm -r ${dirP}
-	cp -r templateReeds $dirP
+		echo ${dirP}
+		rm -r ${dirP}
+		cp -r templateReeds $dirP
 
-	cd ${dirP}
-	./submit_all.sh ${binDIR}
-	cd ..
-
+		cd ${dirP}
+		./submit_all.sh ${binDIR}  || echo "Failed: ${x} " >> Failed.log
+		cd ..
+	fi
 done
